refactor(api-service): split ManagedAgentSetup into focused use cases
The 11-dep god class merged two unrelated entry points (inbound gate vs
OAuth-callback completion) and a third concern (listing OAuth MCPs).
Replace it with two focused use cases and a shared helper.

- `HandleManagedAgentSetupInbound` — gate + park + post/edit setup card
  on inbound. 7 deps, owns `parkAndPostCard` and `resolveStaleSetupCard`.
- `CompleteManagedAgentSetup` — refresh cards or replay parked turns
  after an OAuth callback lands. 11 deps (heavy by nature: cross-entity
  loading + replay), owns the post-callback orchestration.
- `list-oauth-mcps.helper.ts` — free function shared by both use cases;
  `subscriberRepository` / `agentMcpServerRepository` /
  `mcpConnectionRepository` injected by the caller and forwarded.

`setup-card.builder.ts` keeps its existing services-as-params shape (used
by both new use cases). `agent-inbound-handler` and `mcp-oauth-callback`
swap to the new use cases; `usecases/index.ts` registers both. No
behaviour change — replay scope still fans out to all pending
conversations (review item NV-7771 follow-up).
