+++ apps/api/src/app/agents/usecases/mcp-oauth-callback/mcp-oauth-callback.usecase.ts
@@ -35,7 +35,7 @@ import {
 import { MCP_OAUTH_STATE_TTL_MS } from '../generate-mcp-oauth-url/mcp-oauth.constants';
 import { buildMcpOAuthRedirectUri, type McpOAuthState } from '../generate-mcp-oauth-url/mcp-oauth-state';
 import { GetMcpNovuAppCredentials } from '../get-mcp-novu-app-credentials/get-mcp-novu-app-credentials.usecase';
-import { ManagedAgentSetup } from '../managed-agent-setup/managed-agent-setup.usecase';
+import { CompleteManagedAgentSetup } from '../managed-agent-setup/complete-managed-agent-setup.usecase';
 import { ManagedAgentSetupCompleteCommand } from '../managed-agent-setup/managed-agent-setup-complete.command';
 import { SyncAgentMcpServersCommand } from '../sync-agent-mcp-servers/sync-agent-mcp-servers.command';
 import { SyncAgentMcpServers } from '../sync-agent-mcp-servers/sync-agent-mcp-servers.usecase';
@@ -82,7 +82,7 @@ export class McpOAuthCallback {
     private readonly discoveryService: McpOAuthDiscoveryService,
     private readonly syncAgentMcpServers: SyncAgentMcpServers,
     private readonly mcpConnectionVaultService: McpConnectionVaultService,
-    private readonly managedAgentSetup: ManagedAgentSetup,
+    private readonly completeManagedAgentSetup: CompleteManagedAgentSetup,
     private readonly getNovuAppCredentials: GetMcpNovuAppCredentials,
     private readonly logger: PinoLogger
   ) {
@@ -391,7 +391,7 @@ export class McpOAuthCallback {
 
     try {
       // update the agent setup/onboarding card to show the connected MCP
-      await this.managedAgentSetup.handleOAuthConnect(ManagedAgentSetupCompleteCommand.create({ stateData }));
+      await this.completeManagedAgentSetup.execute(ManagedAgentSetupCompleteCommand.create({ stateData }));
     } catch (err) {
       this.logger.warn(
         { err: err instanceof Error ? err.message : String(err), conversationId: stateData.conversationId },
